Module tokio_libtls::prelude

A "prelude" for crates using the tokio-libtls crate.

Re-exports

pub use super::AsyncTls;

Modules

error
Error handling.

Structs

TcpStream
An I/O object representing a TCP stream connected to a remote endpoint.

TlsConfigBuilder
TlsConfigBuilder for TlsConfig.

Enums

Async
Return type of future, indicating whether a value is ready or not.

AsyncSink
The result of an asynchronous attempt to send a value to a sink.

Traits

Future
Trait for types which are a placeholder of a value that may become available at some later point in time.

IntoFuture
Class of types which can be converted into a future.

Sink
A Sink is a value into which other values can be sent, asynchronously.

Stream
A stream of values, not all of which may have been produced yet.

Type Definitions

Poll
Return type of the Future::poll method, indicates whether a future's value is ready or not.

StartSend
Return type of the Sink::start_send method, indicating the outcome of a send attempt. See AsyncSink for more details.